# api/docs/swagger.yaml
basePath: /
definitions:
...
  handler.UserCreateRequest:
    properties:
      max_id:
        example: 123456789
        type: integer
      username:
        example: john_doe
        type: string
    type: object
...
  handler.UserUpdateRequest:
    properties:
      username:
        example: john_doe_updated
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login:
        type: string
      max_id:
        type: integer
      username:
        type: string
    type: object
...
host: 10.155.36.40:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для FarmFocus - системы управления задачами и фермой
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FarmFocus API
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя или обновляет существующего по MaxID.
        При создании также создается статистика и начальные грядки.
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь обновлен
          schema:
            $ref: '#/definitions/model.User'
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать или обновить пользователя
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о текущем аутентифицированном пользователе
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить текущего пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет информацию о текущем пользователе (например, имя пользователя)
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Обновленные данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить данные пользователя
      tags:
      - users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"